+++ Changelog.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+
+## [Unreleased]
+### Fixed
+ - Cursor jumping while typing in specific forms
+
 ## 4.2.3
 ### Fixed
  - Fixed delete action